Added supported sentences for Beitian GPS-module
The Beitian BN-880 appears to use slightly different NEMA sentences. By
adding these sentences and calling the methods which already exist
everything seems to work as expected. This was tested on an ESP32.

The extra exception catching was needed to prevent the program from
crashing in case a "list index out of range" exception was thrown.